Be more compliant with the OpenServiceBroker specification
The Open Service Broker specification specifies that the Metadata fields
in the catalog response are opaque objects. There is a convention that
the specification recommends be adhered to, which matches the structs
defined in this library (ServiceMetadata and ServicePlanMetadata), but
this convention is not mandated by the specification.

Changing the Service and ServicePlan structs to accept an "interface{}"
as the relevant Metadata allows implementors to provide any metadata
information, including the existing structs.