fix: use non-string types for context value keys
The documentation for context.WithValue() says that:
> The provided key must be comparable and should not be of type string
or any other built-in type to avoid collisions between packages using
context.

https://golang.org/pkg/context/#WithValue

Updated to use derived types for context value keys, which are all
exported as constants so that users can import the constant to access
the values.

CorrelationIDKey and RequestIdentityKey were already exported constants, so users
who imported and used these constants should not experience a breaking change.

BREAKING CHANGE