feat!: swap gorilla/mux for go-chi/chi (#246)
gorilla/mux has been archived as a project, and it's preferable to
depend on a maintained HTTP handler. go-chi/chi is the closest fit
for gorilla/mux and appears to be well used and maintained.

An additional breaking change is that go-chi/chi will ignore URL-encoded
slashes for routing purposes, so PUT /v2/service_instances/foo%2Fbar
will treat "foo%2Fbar" as an instance ID and will not try to route to
/v2/service_instances/foo/bar as the gorilla/mux version would have done
by default. Previously the WithEncodedPath() option was used to configure
gorilla/mux to treat "foo%2Fbar" as an instance ID, but as this is now
the default, WithEncodedPath() option has been removed. Treating
"foo%2Fbar" as foo/bar is not an option with go-chi/chi.